VISHU_ELITE_BOT/requirements.txt
MetaTrader5
pandas
numpy
python-dotenv
requests
